chore: refactor(mentor): 콤포넌트 분리 및 연관성 높은 파일명으로 수정
